@@ -22,17 +22,23 @@
   ],
   "minimum-stability": "dev",
   "prefer-stable": true,
+  "config": {
+    "sort-packages": true,
+    "allow-plugins": {
+      "dealerdirect/phpcodesniffer-composer-installer": true
+    }
+  },
   "require": {
-    "php": "^7.3",
+    "php": "^7.4 || ^8.0",
     "zfr/zfr-oauth2-server": "^0.10",
-    "doctrine/orm": "^2.7",
-    "psr/container": "^1.0",
+    "doctrine/orm": "^2.8",
+    "psr/container": "^1.0 || ^2.0",
     "roave/security-advisories": "dev-master"
   },
   "require-dev": {
     "friendsofphp/php-cs-fixer": "^2.1",
-    "phpunit/phpunit": "^8.5",
-    "satooshi/php-coveralls": "~1.0"
+    "laminas/laminas-coding-standard": "^2.4",
+    "phpunit/phpunit": "^9.5"
   },
   "autoload": {
     "psr-4": {
@@ -41,12 +47,22 @@
   },
   "autoload-dev": {
     "psr-4": {
-      "ZfrOAuth2Test\\Server\\Doctrine\\": "tests/src/"
+      "ZfrOAuth2Test\\Server\\Doctrine\\": "test/src/"
     }
   },
   "extra": {
     "branch-alias": {
       "dev-master": "0.3.x-dev"
     }
+  },
+  "scripts": {
+    "check": [
+      "@cs-check",
+      "@test"
+    ],
+    "cs-check": "phpcs",
+    "cs-fix": "phpcbf",
+    "test": "phpunit --colors=always",
+    "test-coverage": "phpunit --colors=always --coverage-clover clover.xml"
   }
 }
