@@ -1,6 +1,7 @@
 <?php
 
 declare(strict_types=1);
+
 /*
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
  * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
@@ -24,9 +25,6 @@ use Psr\Container\ContainerInterface;
 use ZfrOAuth2\Server\Doctrine\Options\DoctrineOptions;
 use ZfrOAuth2\Server\Doctrine\Subscriber\TokenOwnerPkColumnSubscriber;
 
-/**
- * Class TokenOwnerPkColumnSubscriberFactory
- */
 class TokenOwnerPkColumnSubscriberFactory
 {
     public function __invoke(ContainerInterface $container): TokenOwnerPkColumnSubscriber
