@@ -1,6 +1,7 @@
 <?php
 
 declare(strict_types=1);
+
 /*
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
  * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
@@ -20,10 +21,9 @@ declare(strict_types=1);
 
 namespace ZfrOAuth2\Server\Doctrine;
 
+use function array_merge_recursive;
+
 /**
- * Class ModuleConfig
- *
- * @author  Michaël Gallego <mic.gallego@gmail.com>
  * @licence MIT
  */
 class ModuleConfig
