@@ -1,6 +1,7 @@
 <?php
 
 declare(strict_types=1);
+
 /*
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
  * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
@@ -29,26 +30,18 @@ use ZfrOAuth2\Server\Model\Client;
 use ZfrOAuth2\Server\Repository\ClientRepositoryInterface;
 
 /**
- * @author  Michaël Gallego <mic.gallego@gmail.com>
  * @licence MIT
- *
  * @covers  \ZfrOAuth2\Server\Doctrine\Repository\ClientRepository
  */
 class ClientRepositoryTest extends TestCase
 {
-    /**
-     * @var EntityManager|MockObject
-     */
+    /** @var EntityManager|MockObject */
     protected $em;
 
-    /**
-     * @var ClassMetadata|MockObject
-     */
+    /** @var ClassMetadata|MockObject */
     protected $meta;
 
-    /**
-     * @var ClientRepository
-     */
+    /** @var ClientRepository */
     protected $repository;
 
     public function setUp(): void
@@ -67,11 +60,11 @@ class ClientRepositoryTest extends TestCase
     {
         $client = $this->createMock(Client::class);
 
-        $this->em->expects($this->at(0))
+        $this->em->expects($this->once())
             ->method('persist')
             ->with($client);
 
-        $this->em->expects($this->at(1))
+        $this->em->expects($this->once())
             ->method('flush')
             ->with($client);
 
@@ -82,7 +75,7 @@ class ClientRepositoryTest extends TestCase
 
     public function testFindById(): void
     {
-        $this->em->expects($this->at(0))
+        $this->em->expects($this->once())
             ->method('find')
             ->willReturn(null);
 
@@ -93,7 +86,7 @@ class ClientRepositoryTest extends TestCase
 
     public function testIdExistsTrue(): void
     {
-        $this->em->expects($this->at(0))
+        $this->em->expects($this->once())
             ->method('find')
             ->willReturn($this->createMock(Client::class));
 
@@ -104,7 +97,7 @@ class ClientRepositoryTest extends TestCase
 
     public function testIdExistsFalse(): void
     {
-        $this->em->expects($this->at(0))
+        $this->em->expects($this->once())
             ->method('find')
             ->willReturn(null);
 
